requirements.txt
mysql-connector-python==v8.0.31